requirements.txt
selenium==4.18.1
undetected-chromedriver==3.5.5
webdriver-manager==4.0.1

groq==0.4.2
pydantic==2.6.3

python-dotenv==1.0.1
setuptools 

pandas==2.2.1